+++ instrumentation/servlet/servlet-common/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/servlet/common/async/AsyncContextInstrumentation.java
@@ -3,16 +3,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.servlet.v5_0.dispatcher;
+package io.opentelemetry.javaagent.instrumentation.servlet.common.async;
 
 import static io.opentelemetry.javaagent.tooling.bytebuddy.matcher.AgentElementMatchers.implementsInterface;
 import static io.opentelemetry.javaagent.tooling.bytebuddy.matcher.ClassLoaderMatcher.hasClassesNamed;
-import static io.opentelemetry.javaagent.tooling.bytebuddy.matcher.NameMatchers.namedOneOf;
 import static java.util.Collections.singletonMap;
+import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
-import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 import io.opentelemetry.javaagent.tooling.TypeInstrumentation;
 import java.util.Map;
@@ -20,33 +18,27 @@ import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
-public class RequestDispatcherInstrumentation implements TypeInstrumentation {
+public class AsyncContextInstrumentation implements TypeInstrumentation {
   private final String basePackageName;
   private final String adviceClassName;
 
-  public RequestDispatcherInstrumentation(String basePackageName, String adviceClassName) {
+  public AsyncContextInstrumentation(String basePackageName, String adviceClassName) {
     this.basePackageName = basePackageName;
     this.adviceClassName = adviceClassName;
   }
 
   @Override
   public ElementMatcher<ClassLoader> classLoaderOptimization() {
-    return hasClassesNamed(basePackageName + ".RequestDispatcher");
+    return hasClassesNamed(basePackageName + ".AsyncContext");
   }
 
   @Override
   public ElementMatcher<TypeDescription> typeMatcher() {
-    return implementsInterface(named(basePackageName + ".RequestDispatcher"));
+    return implementsInterface(named(basePackageName + ".AsyncContext"));
   }
 
   @Override
   public Map<? extends ElementMatcher<? super MethodDescription>, String> transformers() {
-    return singletonMap(
-        namedOneOf("forward", "include")
-            .and(takesArguments(2))
-            .and(takesArgument(0, named(basePackageName + ".ServletRequest")))
-            .and(takesArgument(1, named(basePackageName + ".ServletResponse")))
-            .and(isPublic()),
-        adviceClassName);
+    return singletonMap(isMethod().and(isPublic()).and(named("dispatch")), adviceClassName);
   }
 }
