+++ instrumentation/servlet/servlet-common/library/src/main/java/io/opentelemetry/instrumentation/servlet/ServletAccessor.java
@@ -7,6 +7,15 @@ package io.opentelemetry.instrumentation.servlet;
 
 import java.security.Principal;
 
+/**
+ * This interface is used to access methods of HttpServletRequest and HttpServletResponse classes in
+ * shared code that is used for both jakarta.servlet and javax.servlet versions of those classes. A
+ * wrapper class with extra information attached may be used as well in cases where the class itself
+ * does not provide some field (such as response status for Servlet API 2.2).
+ *
+ * @param <RequestT> HttpServletRequest class (or a wrapper)
+ * @param <ResponseT> HttpServletResponse class (or a wrapper)
+ */
 public interface ServletAccessor<RequestT, ResponseT> {
   String getRequestContextPath(RequestT request);
