+++ javaagent-api/src/main/java/io/opentelemetry/javaagent/instrumentation/api/CallDepth.java
@@ -28,7 +28,7 @@ public final class CallDepth {
     return depth;
   }
 
-  void reset() {
+  public void reset() {
     depth = 0;
   }
 }
